Separated project cards into two components and added support for different views of the cards
